machine-learning/requirements.txt


numpy
pandas
scikit-learn
concrete-ml
tenseal
torch
psutil
tracemalloc    
matplotlib
seaborn
gc                      
dataclasses             
typing                 
jupyter
ipython
cryptography
protobuf
pybind11

